Backend/equipment/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==41.0.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.2.5
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
flatbuffers==23.3.3
fonttools==4.39.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.31.1
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jax==0.4.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.13.1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.10.1
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pipenv==2023.6.2
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-localtunnel==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-cpu==2.12.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
